requirements.txt
# Telegram Bot Library
python-telegram-bot[job-queue]==21.0.1

# PostgreSQL Database Driver
psycopg[binary]==3.1.18
psycopg-pool==3.2.2

# Web Server & Templates
aiohttp==3.9.1
aiohttp-jinja2==1.6
Jinja2>=3.1.2
